requirements.txt
cython
git+https://github.com/cocodataset/cocoapi.git#subdirectory=PythonAPI&egg=pycocotools
submitit
git+https://github.com/cocodataset/panopticapi.git#egg=panopticapi
scipy
termcolor
addict
yapf
timm
Pillow==10.1.0
numpy
pandas
opencv-python
protobuf==5.26.1
tqdm
scikit-learn==1.3.2
matplotlib